backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.2
mysqlclient==2.2.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==5.0.0
stripe==8.4.0

